Refactor main Helper class
