+++ composer/helpers/bin/run
@@ -13,7 +13,7 @@ $request = json_decode(file_get_contents('php://stdin'), true);
 
 // Increase the default memory limit the same way Composer does (but clearer)
 if (function_exists('ini_set')) {
-    $memoryInBytes = static function ($value) {
+    $memoryInBytes = static function (string $value): int {
         $unit = strtolower(substr($value, -1, 1));
         $value = (int) $value;
         if ($unit == 'g') {
@@ -29,7 +29,7 @@ if (function_exists('ini_set')) {
 
     $memoryLimit = trim(ini_get('memory_limit'));
     // Increase memory_limit if it is lower than 1900MB
-    if ($memoryLimit != -1 && $memoryInBytes($memoryLimit) < 1024 * 1024 * 1900) {
+    if ($memoryLimit !== "" && $memoryInBytes($memoryLimit) < 1024 * 1024 * 1900) {
         @ini_set('memory_limit', '1900M');
     }
 
@@ -45,38 +45,43 @@ date_default_timezone_set('Europe/London');
 // This storage is freed on error (case of allowed memory exhausted)
 $memory = str_repeat('*', 1024 * 1024);
 
+function output(array $content): void {
+    fwrite(STDOUT, json_encode($content));
+}
+
 register_shutdown_function(function (): void {
     global $memory;
     $memory = null;
     $error = error_get_last();
     if (null !== $error) {
-        fwrite(STDOUT, json_encode(['error' => $error['message']]));
+        output(['error' => $error['message']]);
     }
 });
 
 try {
     switch ($request['function']) {
         case 'update':
-            $updatedFiles = Updater::update($request['args']);
-            fwrite(STDOUT, json_encode(['result' => $updatedFiles]));
+            [$workingDirectory, $dependencyName, $dependencyVersion, $gitCredentials, $registryCredentials] = $request['args'];
+            $updatedFiles = Helper::update($workingDirectory, $dependencyName, $gitCredentials, $registryCredentials);
+            output(['result' => $updatedFiles]);
             error_clear_last();
             break;
         case 'get_latest_resolvable_version':
-            $latestVersion = UpdateChecker::getLatestResolvableVersion($request['args']);
-            fwrite(STDOUT, json_encode(['result' => $latestVersion]));
+            $latestVersion = Helper::getLatestResolvableVersion(...$request['args']);
+            output(['result' => $latestVersion]);
             error_clear_last();
             break;
         case 'get_content_hash':
-            $content_hash = Hasher::getContentHash($request['args']);
-            fwrite(STDOUT, json_encode(['result' => $content_hash]));
+            $content_hash = Helper::getContentHash(...$request['args']);
+            output(['result' => $content_hash]);
             error_clear_last();
             break;
         default:
-            fwrite(STDOUT, json_encode(['error' => "Invalid function {$request['function']}"]));
+            output(['error' => "Invalid function {$request['function']}"]);
             exit(1);
     }
 } catch (\Exception $e) {
-    fwrite(STDOUT, json_encode(['error' => $e->getMessage()]));
+    output(['error' => $e->getMessage()]);
     error_clear_last();
     exit(1);
 }
