+++ composer/helpers/src/Installer.php
@@ -1,0 +1,72 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Dependabot\Composer;
+
+use Composer\Composer;
+use Composer\Installer as Base;
+use Composer\IO\IOInterface;
+
+class Installer extends Base
+{
+    public function __construct(
+        IOInterface $io,
+        Composer $composer,
+        string $dependencyName,
+        array $gitCredentials,
+        array $registryCredentials,
+        ?InstallationManager $installationManager = null
+    ) {
+        $httpBasicCredentials = [];
+
+        foreach ($gitCredentials as $credentials) {
+            $httpBasicCredentials[$credentials['host']] = [
+                'username' => $credentials['username'],
+                'password' => $credentials['password'],
+            ];
+        }
+
+        foreach ($registryCredentials as $credentials) {
+            $httpBasicCredentials[$credentials['registry']] = [
+                'username' => $credentials['username'],
+                'password' => $credentials['password'],
+            ];
+        }
+
+        $config = $composer->getConfig();
+
+        if (0 < count($httpBasicCredentials)) {
+            $config->merge([
+                'config' => [
+                    'http-basic' => $httpBasicCredentials,
+                ],
+            ]);
+
+            $io->loadConfiguration($config);
+        }
+
+        parent::__construct(
+            $io,
+            $config,
+            $composer->getPackage(),
+            $composer->getDownloadManager(),
+            $composer->getRepositoryManager(),
+            $composer->getLocker(),
+            $installationManager ?: $composer->getInstallationManager(),
+            $composer->getEventDispatcher(),
+            $composer->getAutoloadGenerator()
+        );
+
+        // For all potential options, see UpdateCommand in composer
+        $this
+            ->setUpdate(true)
+            ->setDevMode(true)
+            ->setUpdateWhitelist([$dependencyName])
+            ->setWhitelistTransitiveDependencies(true)
+            ->setExecuteOperations(false)
+            ->setDumpAutoloader(false)
+            ->setRunScripts(false)
+            ->setIgnorePlatformRequirements(false);
+    }
+}
