--- composer/helpers/src/Updater.php
@@ -1,102 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Dependabot\Composer;
-
-use Composer\Factory;
-use Composer\Installer;
-
-class Updater
-{
-    /**
-     * @param array $args
-     *
-     * @throws \RuntimeException
-     *
-     * @return array
-     */
-    public static function update(array $args): array
-    {
-        [$workingDirectory, $dependencyName, $dependencyVersion, $gitCredentials, $registryCredentials] = $args;
-
-        // Change working directory to the one provided, this ensures that we
-        // install dependencies into the working dir, rather than a vendor folder
-        // in the root of the project
-        $originalDir = getcwd();
-
-        if (!is_string($originalDir)) {
-            throw new \RuntimeException('Failed determining the current working directory.');
-        }
-
-        chdir($workingDirectory);
-
-        $io = new ExceptionIO();
-        $composer = Factory::create($io);
-        $config = $composer->getConfig();
-        $httpBasicCredentials = [];
-
-        $pm = new DependabotPluginManager($io, $composer, null, false);
-        $composer->setPluginManager($pm);
-        $pm->loadInstalledPlugins();
-
-        foreach ($gitCredentials as &$cred) {
-            $httpBasicCredentials[$cred['host']] = [
-                'username' => $cred['username'],
-                'password' => $cred['password'],
-            ];
-        }
-
-        foreach ($registryCredentials as &$cred) {
-            $httpBasicCredentials[$cred['registry']] = [
-                'username' => $cred['username'],
-                'password' => $cred['password'],
-            ];
-        }
-
-        if ($httpBasicCredentials) {
-            $config->merge(
-                [
-                    'config' => [
-                        'http-basic' => $httpBasicCredentials,
-                    ],
-                ]
-            );
-            $io->loadConfiguration($config);
-        }
-
-        $install = new Installer(
-            $io,
-            $config,
-            $composer->getPackage(),
-            $composer->getDownloadManager(),
-            $composer->getRepositoryManager(),
-            $composer->getLocker(),
-            $composer->getInstallationManager(),
-            $composer->getEventDispatcher(),
-            $composer->getAutoloadGenerator()
-        );
-
-        // For all potential options, see UpdateCommand in composer
-        $install
-            ->setWriteLock(true)
-            ->setUpdate(true)
-            ->setDevMode(true)
-            ->setUpdateWhitelist([$dependencyName])
-            ->setWhitelistTransitiveDependencies(true)
-            ->setExecuteOperations(false)
-            ->setDumpAutoloader(false)
-            ->setRunScripts(false)
-            ->setIgnorePlatformRequirements(false);
-
-        $install->run();
-
-        $result = [
-            'composer.lock' => file_get_contents('composer.lock'),
-        ];
-
-        chdir($originalDir);
-
-        return $result;
-    }
-}
